Move bstats classes to their own package.
