+++ src/main/java/org/betonquest/betonquest/BetonQuest.java
@@ -186,9 +186,9 @@ import org.betonquest.betonquest.objectives.SmeltingObjective;
 import org.betonquest.betonquest.objectives.StepObjective;
 import org.betonquest.betonquest.objectives.TameObjective;
 import org.betonquest.betonquest.objectives.VariableObjective;
-import org.betonquest.betonquest.utils.BStatsMetrics;
-import org.betonquest.betonquest.utils.CompositeInstructionMetricsSupplier;
-import org.betonquest.betonquest.utils.InstructionMetricsSupplier;
+import org.betonquest.betonquest.bstats.BStatsMetrics;
+import org.betonquest.betonquest.bstats.CompositeInstructionMetricsSupplier;
+import org.betonquest.betonquest.bstats.InstructionMetricsSupplier;
 import org.betonquest.betonquest.utils.PlayerConverter;
 import org.betonquest.betonquest.utils.Utils;
 import org.betonquest.betonquest.utils.versioning.Updater;
